Move SSE related code to new file
no functional changes
